Update 11-Decomp.md
I moved it to the bottom, however as far as the "extra parenthesis", the link itself has an open parenthesis character '(' and according to GitHub it looks like an open parenthesis considering that you need to put the link in parenthesis. If I were to remove it, it would cause an error.